+++ R/warpHorizons.R
@@ -7,9 +7,19 @@
 #' 
 #' @description This function applies a warping factor to the horizons of a single-profile `SoilProfileCollection` object. Warping values >1 will inflate horizon thickness, values <1 will deflate horizon thickness.
 #'
-#' @param x a `SoilProfileCollection` object
-#' @param fact numeric or character; warping factor specified as a single numeric value, vector of numeric values (length = nrow(x)), or column name of a horizon level attribute containing numeric values
+#' @param x a `SoilProfileCollection` object with a single soil profile
+#' 
+#' @param fact numeric or character; warping factor specified as a single numeric value, vector of numeric values (length = `nrow(x)`), or column name of a horizon level attribute containing numeric values
+#' 
+#' @param scaleTo numeric, target depth resulting from rescaling all horizon thickness values, due to rounding the actual depth may be within +/- 1 depth unit
+#' 
+#' @param soilDepthFun function for computing soil depth, either: 
+#' 
+#'   * `estimateSoilDepth()`: "soil depth" is defined as depth to contact, requires that `hzdesgnname()` be set, see [estimateSoilDepth()] for details
+#'   * `max()`: "soil depth" is defined as the bottom depth of the soil profile
+#' 
 #' @param updateProfileID logical; modify profile IDs
+#' 
 #' @param suffix character; suffix added to profile IDs when `updateProfileID = TRUE`
 #' 
 #' @author D.E. Beaudette and S.W. Salley
@@ -50,7 +60,7 @@
 #'   cex.names = 0.8, 
 #'   width = 0.2, 
 #'   max.depth = 150, 
-#'   depth.axis = list(line = -3), 
+#'   depth.axis = FALSE, 
 #'   y.offset = .yoff
 #' )
 #' 
@@ -66,7 +76,7 @@
 #' .template <- c(
 #' 'P1:AAA|BwBwBwBw|CCCCCCC|CdCdCdCd',
 #' 'P2:ApAp|AA|E|BhsBhs|Bw1Bw1|CCCCC',
-#' 'P3: A|Bt1Bt1Bt1|Bt2Bt2Bt2|Bt3|Cr|RRRRR'
+#' 'P3:A|Bt1Bt1Bt1|Bt2Bt2Bt2|Bt3|Cr|RRRRR'
 #' )
 #' 
 #' # each horizon label is '10' depth-units (default)
@@ -95,15 +105,50 @@
 #'   depth.axis = list(line = -2)
 #' )
 #' 
+#' # rescale all profiles to 100cm soil depth (depth to contact)
+#' .template <- c(
+#'   'P1:AAA|BwBwBwBw|CCCCCCC|CdCdCdCd',
+#'   'P2:ApAp|AA|E|BhsBhs|Bw1Bw1|CCCCC',
+#'   'P3:A|Bt1Bt1Bt1|Bt2Bt2Bt2|Bt3|Cr|RRRRR',
+#'   'P4:AAAAA|CCC|RRRRRR'
+#' )
+#' 
+#' # each horizon label is '10' depth-units (default)
+#' s <- quickSPC(.template)
+#' 
+#' # warp horizons by profile, result is a list of SPCs
+#' w <- profileApply(s, FUN = function(i) {
+#'   warpHorizons(i, scaleTo = 100, soilDepthFun = estimateSoilDepth)
+#' })
+#' 
+#' # flatten list -> SoilProfileCollection
+#' w <- combine(w)
+#' 
+#' # combine with original SPC
+#' x <- combine(s, w)
+#' 
+#' # highlight "contact"
+#' x$color <- rep(grey(0.9), times = nrow(x))
+#' x$color[grep('R|Cr|Cd', x$name)] <- 'royalblue'
+#' 
+#' # sketches
+#' par(mar = c(0.5, 0, 0, 2.5))
+#' plotSPC(
+#'   x,
+#'   color = 'color',
+#'   name.style = 'center-center',
+#'   cex.names = 0.8,
+#'   width = 0.3,
+#'   max.depth = 165,
+#'   depth.axis = list(line = -2)
+#' )
+#' 
+#' abline(h = 100, lty = 3)
 #'
-warpHorizons <- function(x, fact, updateProfileID = TRUE, suffix = '-w') {
+warpHorizons <- function(x, fact = NULL, scaleTo = NULL, soilDepthFun = estimateSoilDepth, updateProfileID = TRUE, suffix = '-w') {
   
   ## TODO: 
   ## * vectorize over profiles, and make more efficient
-  ## * support the more general, Ax + c affine transformation for 1D geometry
-  ##   -> https://www.cs.utexas.edu/users/fussell/courses/cs384g-fall2011/lectures/lecture07-Affine.pdf
-  
-
   
   # extract parts and pieces of the SPC
   .h <- horizons(x)
@@ -111,23 +156,40 @@ warpHorizons <- function(x, fact, updateProfileID = TRUE, suffix = '-w') {
   .n <- nrow(.h)
   
   # sanity check: fact should be:
+  # * NULL, not specified in favor of scaleTo
   # * numeric, length == 1, used by all horizons
   # * numeric, length == nrow(x), each horizon has its own factor
   # * hz column name
   
-  if(inherits(fact, 'character')) {
-    fact <- x[[fact]]
-    if(is.null(fact)) {
-      stop('fact must name a column in horizons')
+  if(!is.null(fact)) {
+    if(inherits(fact, 'character')) {
+      fact <- x[[fact]]
+      if(is.null(fact)) {
+        stop('fact must name a column in horizons', call. = FALSE)
+      }
+    } else if(inherits(fact, 'numeric')) {
+      
+      if(length(fact) > 1 && length(fact) != .n) {
+        stop('fact must be length 1 or nrow(x)', call. = FALSE)
+      }
+      
+    } else {
+      stop('fact must be either character vector or numeric', call. = FALSE)
     }
-  } else if(inherits(fact, 'numeric')) {
+  } else if(!is.null(scaleTo)) {
     
-    if(length(fact) > 1 && length(fact) != .n) {
-      stop('fact must be length 1 or nrow(x)')
+    if(length(scaleTo) > 1) {
+      stop('scaleTo should have length of 1')
     }
+    
+    # compute fact from scaleTo and soil depth
+    # this requires
+    fact <- (1 / soilDepthFun(x)) * scaleTo
     
   } else {
-    stop('fact must be either character vector or numeric')
+    # identity
+    message('no scaling factor provided, using `fact = 1`')
+    fact <- 1
   }
   
   # warping is applied to horizon thickness